Update README.md
This is a react ToDo Application with local storage. This application enables you to add a task and the date that task is expected to be complete. On Clicking the submit the button, the task is stored in the local storage togther with the expected date and the status of the progress.